requirements.txt
# LangChain 核心依赖
langchain
langchain-community
langchain-deepseek
langchain-anthropic
langgraph

# 数据解析
lxml

# 环境变量
python-dotenv

# 向量数据库
chromadb
onnxruntime

# 数据处理
pandas
numpy<2.0.0

# 数据库驱动
sqlalchemy
pymysql
oracledb

# FastAPI Web 框架
fastapi
uvicorn[standard]
pydantic
